+++ src/cache/types.go
@@ -1,6 +1,6 @@
 package cache
 
 type cachedData struct {
-	Positions map[string]int `json:"positions"`
-	Channels  []string       `json:"channels"`
+	Positions map[string]uint64 `json:"positions"`
+	Channels  []string          `json:"channels"`
 }
